feat(swc): integrate token(), include stylesheets
