+++ 2nd-gen/packages/swc/stylesheets/swc.css
@@ -1,0 +1,47 @@
+/**
+ * Copyright 2025 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
+@import url('tokens.css');
+
+/* TODO: update to expose/use values similar to previous setup */
+:root {
+    font-family: var(--swc-sans-serif-font-family), 'Adobe Clean', sans-serif;
+    font-size: var(--swc-font-size-200);
+    font-style: var(--swc-default-font-style);
+}
+
+/* Theme scaling setup */
+:root,
+.spectrum-theme,
+.spectrum-theme--sizeM {
+    --spectrum-theme-size: var(--spectrum-theme--sizeM,);
+    --spectrum-theme--sizeM: var(--spectrum-theme-size,);
+    --spectrum-theme--sizeL: var(--spectrum-theme-size,);
+}
+
+.spectrum-theme--sizeL {
+    --spectrum-theme-size: var(--spectrum-theme--sizeL);
+}
+
+/* Theme colors setup */
+:root,
+.spectrum-theme--light {
+    color-scheme: light;
+}
+
+.spectrum-theme--dark {
+    color-scheme: dark;
+}
+
+.spectrum-theme--adaptive {
+    color-scheme: light dark;
+}
